requirements.txt
asgiref==3.8.1
coverage==7.6.8
Django==5.1.3
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
git-filter-repo==2.45.0
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.2
tzdata==2024.2
